+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 0.18.0
 
+* Add possibility to set Java System Properties for User Operator and Topic Operator via `Kafka` CR.
 * Make it possible to configure PodManagementPolicy for StatefulSets
 * Update build system to use `yq` version 3 (https://github.com/mikefarah/yq)